a few bug fixes and moved auth.js from service folder to src.
