+++ package.json
@@ -18,7 +18,8 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "concurrently": "^8.2.2"
+    "concurrently": "^8.2.2",
+    "graphql-tag": "^2.12.6"
   },
   "devDependencies": {
     "wait-on": "^7.2.0"
