+++ server/src/models/User.js
@@ -3,7 +3,7 @@ import bcrypt from "bcrypt";
 
 // User Schema
 const userSchema = new mongoose.Schema({
-  name: String,
+  username: String,
   email: { type: String, unique: true },
   password: {
     type: String,
